Split IChannel file into its constituent types
