+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/ChannelUtilities.cs
@@ -3,8 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Runtime.CompilerServices;
 
 namespace System.Threading.Tasks.Channels
 {
